+++ lib/dart_failure_handler.dart
@@ -1,6 +1,0 @@
-library;
-
-export 'src/either.dart';
-export 'src/failure.dart';
-export 'src/error_mapper.dart';
-export 'src/repository_mixin.dart';