+++ lib/failure_kit.dart
@@ -1,0 +1,4 @@
+export 'src/either.dart';
+export 'src/failure.dart';
+export 'src/failure_mapper.dart';
+export 'src/failure_guard.dart';